Move uikit out of shared
