+++ frontend/src/assets/scss/index.scss
@@ -34,4 +34,4 @@
 @import 'form/switch';
 
 @import '../../../config/styles';
-@import '../../shared/ui-kit';
+@import '../../ui-kit';
