+++ frontend/src/ui-kit/card/card.scss
