+++ frontend/src/ui-kit/checkbox/checkbox.scss
