+++ frontend/src/ui-kit/checkbox/types/CheckboxSize.ts
