+++ frontend/src/ui-kit/field-message/FieldMessage.vue
@@ -14,9 +14,9 @@
 </template>
 
 <script setup lang="ts">
-import { FieldMessageType } from '@/shared/ui-kit/field-message/types/FieldMessageType';
+import { FieldMessageType } from '@/ui-kit/field-message/types/FieldMessageType';
 import { computed } from 'vue';
-import { fieldMessageTypeData } from '@/shared/ui-kit/field-message/constants/fieldMessageTypeData';
+import { fieldMessageTypeData } from '@/ui-kit/field-message/constants/fieldMessageTypeData';
 
 const props = withDefaults(defineProps<{
   type?: FieldMessageType,
