+++ frontend/src/ui-kit/input-password/InputPassword.stories.ts
