+++ frontend/src/ui-kit/switch/switch.scss
