+++ frontend/src/ui-kit/switch/types/SwitchSize.ts
